refactor(types): port metadata types to explicit imports
