+++ src/api/Metadata.js
@@ -29,7 +29,13 @@ import {
 } from '../constants';
 
 import type { FetchOptions, ElementsErrorCallback } from '../common/types/api';
-import type { MetadataTemplateSchemaResponse } from '../common/types/metadata';
+import type {
+    MetadataTemplateSchemaResponse,
+    MetadataTemplate,
+    MetadataInstanceV2,
+    MetadataEditor,
+    MetadataFields,
+} from '../common/types/metadata';
 
 class Metadata extends File {
     /**
