+++ src/elements/content-sidebar/Sidebar.js
@@ -23,6 +23,7 @@ import type { DetailsSidebarProps } from './DetailsSidebar';
 import type { MetadataSidebarProps } from './MetadataSidebar';
 import type { VersionsSidebarProps } from './versions';
 import type { AdditionalSidebarTab } from './flowTypes';
+import type { MetadataEditor } from '../../common/types/metadata';
 
 type Props = {
     activitySidebarProps: ActivitySidebarProps,
