+++ src/features/metadata-based-view/MetadataBasedItemList.js
@@ -20,6 +20,7 @@ import messages from '../../elements/common/messages';
 import './MetadataBasedItemList.scss';
 
 import type { MetadataColumnConfig, MetadataColumnsToShow } from '../../common/types/metadataQueries';
+import type { MetadataFieldValue } from '../../common/types/metadata';
 
 import {
     CANCEL_ICON_TYPE,
