+++ src/features/metadata-instance-editor/MetadataInstanceEditor.js
@@ -6,6 +6,7 @@ import ScrollWrapper from '../../components/scroll-wrapper';
 import Header from './Header';
 import Instances from './Instances';
 import EmptyContent from './EmptyContent';
+import type { MetadataEditor, MetadataTemplate, MetadataCascadingPolicyData } from '../../common/types/metadata';
 import './MetadataInstanceEditor.scss';
 
 type Props = {
