+++ src/features/metadata-instance-editor/TemplatedInstance.js
@@ -5,6 +5,7 @@ import { FormattedMessage } from 'react-intl';
 import Field from './fields/Field';
 import messages from './messages';
 import { isHidden } from './metadataUtil';
+import type { MetadataFields, MetadataTemplate, MetadataFieldValue } from '../../common/types/metadata';
 import './TemplatedInstance.scss';
 
 type Props = {
